LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/plaintext-only/plaintext-only-in-designMode-expected.txt

AB


FAIL contenteditable=plaintext-only in the design mode should not prevent the style edit assert_equals: expected "<div contenteditable=\"plaintext-only\"><b>ABC</b></div>" but got "<div contenteditable=\"plaintext-only\">ABC</div>"
PASS contenteditable=plaintext-only in the design mode should not suppress style of inserting HTML
FAIL contenteditable=plaintext-only in the design mode should not cause inserting line break at typing Enter assert_equals: expected "<div contenteditable=\"plaintext-only\"><p>A</p><p>B</p></div>" but got "<div contenteditable=\"plaintext-only\"><p>AB</p></div>"
